+++ jni/application/onscripter/backtrace.h
@@ -1,7 +1,7 @@
 #ifndef __BACKTRACE_H__
 #define __BACKTRACE_H__
 
-unsigned int get_backtrace(char** out, unsigned int buffSize);
+size_t get_backtrace(char** out, size_t buffSize);
 
 void print_backtrace(const char* label = NULL);
